+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Back to [Readme](README.md).
 
+## [3.13.0] - UNRELEASED
+
+### Added
+
+* New `expandErrorMessages` option to expand or collapse error messages by default on scenario overview pages (default: false) (#391)
+
+### Changed
+
+* Minimum required Java version is now 17
+* More readable data table formatting
+* More consistent error message formatting
+
 ## [3.12.0] - 2025-11-03
 
 ### Added
@@ -1044,6 +1056,8 @@ the core component is now the reporting engine that is the base for other forms 
 ## [0.0.2] - 2017-11-14
 
 Initial project version on GitHub and Maven Central.
+
+[3.13.0]: https://github.com/trivago/cluecumber-report-plugin/tree/v3.13.0
 
 [3.12.0]: https://github.com/trivago/cluecumber-report-plugin/tree/v3.12.0